seperated our code into different components
