+++ package.json
@@ -7,7 +7,8 @@
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
     "react-firebase-hooks": "^3.0.4",
-    "react-scripts": "4.0.1"
+    "react-scripts": "4.0.1",
+    "react-uuid": "^1.0.2"
   },
   "devDependencies": {
     "@testing-library/jest-dom": "^5.11.6",
