--- src/components/GroceryComponent.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const GroceryComponent = () => {
-  return <div>GROCERIES</div>;
-};
-
-export default GroceryComponent;
